requirements.txt
asgiref==3.6.0
attrs==22.2.0
awsebcli==3.20.3
bcrypt==4.0.1
blessed==1.19.1
botocore==1.23.54
cached-property==1.5.2
cement==2.8.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.0.12
cloudinary==1.28.1
colorama==0.4.3
cryptography==39.0.0
distlib==0.3.6
dj3-cloudinary-storage==0.0.6
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
dj-database-url==1.2.0
Django==4.0.1
django-cors-headers==3.11.0
django-rest-knox==4.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
filelock==3.9.0
future==0.16.0
gunicorn==20.1.0
idna==3.4
Jinja2==3.0.3
jmespath==0.10.0
jsonschema==3.2.0
Markdown==3.3.6
MarkupSafe==2.0.1
mock==5.0.1
paramiko==2.12.0
pathspec==0.9.0
pickleshare==0.7.5
Pillow==9.0.0
prompt-toolkit==3.0.21
psycopg2==2.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
platformdirs==2.6.2
pycparser==2.21
PyJWT==2.3.0
PyNaCl==1.5.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.7.1
PyYAML==5.4.1
requests==2.26.0
semantic-version==2.8.5
six==1.14.0
sqlparse==0.4.3
termcolor==1.1.0
texttable==1.6.7
urllib3==1.26.14
virtualenv==20.17.1
wcwidth==0.1.9
websocket-client==0.59.0
